\name{DImodels-internal}
\alias{proflik_theta}
\alias{test_autoDI}
\alias{AICsel_autoDI}
\alias{AICcsel_autoDI}
\alias{BICsel_autoDI}
\alias{BICcsel_autoDI}
\alias{namesub_autoDI}
\alias{namesub_DI}
\alias{allmodels_autoDI}
\alias{AIC2}
\alias{AICc}
\alias{AICc.default}
\alias{BIC2}
\alias{BICc}
\alias{BICc.default}
\alias{DI_matrix_check}
\alias{get_P_indices}
\alias{get_theta_cols}
\alias{theta_index_matching}
\alias{reftest_autoDI}
\alias{get_community}
\alias{DI_reference}
\alias{anova.DI}
\alias{anovaDIglm}
\alias{anova_glmlist}
\alias{autoDI_step0}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal Functions
}
\description{
Internal functions used within the \code{DI} and \code{autoDI} functions.
}
\author{
Rafael A. Moral, John Connolly and Caroline Brophy
}